:truck: Move tracing functions to infrastructure
