+++ marketplace-core/src/main.rs
@@ -1,48 +1,13 @@
-mod datadog_event_format;
-mod marketplace_tracing;
 use anyhow::Result;
 use dotenv::dotenv;
+use marketplace_infrastructure::tracing::setup_tracing;
 
 extern crate rocket;
-
-#[tracing::instrument]
-fn foo(bar: String) -> String {
-	// tracing::warn_span!("New WARN span?");
-	// tracing::warn!("This is my WARN trace!");
-	// let a = get_tracing_id();
-	// let b = get_tracing_span_id();
-	// log::warn!("BBBBBBBBBBB! TRACE {}, SPAN {}", a, b);
-	log::error!("to");
-	fee(bar) + " smith"
-}
-
-#[tracing::instrument]
-fn fee(bar: String) -> String {
-	// tracing::warn_span!("New WARN span?");
-	// tracing::warn!("This is my WARN trace!");
-	// let a = get_tracing_id();
-	// let b = get_tracing_span_id();
-	// log::warn!("BBBBBBBBBBB! TRACE {}, SPAN {}", a, b);
-	log::info!("kyo");
-	tracing::event!(
-		tracing::Level::INFO,
-		msg = "message",
-		answer = 42,
-		question = "life, the universe, and everything",
-		"what?"
-	);
-	bar + " PhD"
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
 	dotenv().ok();
 
-	marketplace_tracing::setup_tracing();
-	//let _global_logger_guard = logger::set_global_logger(create_root_logger());
-
-	tracing::info!("This is my INFO trace!");
-	log::info!("This is my INFO log!");
-	foo("john".to_string());
+	setup_tracing();
 	marketplace_core::main().await
 }
