+++ marketplace-indexer/src/main.rs
@@ -1,11 +1,11 @@
 use anyhow::Result;
 use dotenv::dotenv;
-use marketplace_infrastructure::logger;
+use marketplace_infrastructure::tracing::setup_tracing;
 
 #[tokio::main]
 async fn main() -> Result<()> {
 	dotenv().ok();
-	logger::set_default_global_logger().cancel_reset();
+	setup_tracing();
 
 	marketplace_indexer::main().await
 }
