+++ marketplace-infrastructure/Cargo.toml
@@ -16,8 +16,8 @@ with_infrastructure_tests = []
 
 # Async
 async-trait = "0.1.56"
-futures = { version = "0.3.21", features = ["alloc"] }
-tokio = { version = "1.18.2", features = ["full"] }
+futures = {version = "0.3.21", features = ["alloc"]}
+tokio = {version = "1.18.2", features = ["full"]}
 tokio-stream = "0.1.10"
 
 # RabbitMQ
@@ -25,34 +25,34 @@ lapin = "2.1.1"
 tokio-executor-trait = "2.1.0"
 
 # Database
-diesel = { version = "1.4.8", default-features = false, features = [
-	"postgres",
-	"serde_json",
-	"uuidv07",
-	"chrono",
-] }
+diesel = {version = "1.4.8", default-features = false, features = [
+  "postgres",
+  "serde_json",
+  "uuidv07",
+  "chrono",
+]}
 diesel_migrations = "1.4.0"
 r2d2 = "0.8.10"
 r2d2-diesel = "1.0.0"
-uuid = { version = "0.8.2", default_features = false, features = [
-	"v4",
-	"serde",
-] }
+uuid = {version = "0.8.2", default_features = false, features = [
+  "v4",
+  "serde",
+]}
 
 # Github
 octocrab = "0.16.0"
 
 # Http Server
-rocket = { version = "0.5.0-rc.2", features = ["json", "uuid"] }
+rocket = {version = "0.5.0-rc.2", features = ["json", "uuid"]}
 
 # Http Client
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = {version = "0.11", features = ["json"]}
 
 # Starknet
-starknet = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet = {git = "https://github.com/xJonathanLEI/starknet-rs"}
 
 # Math types
-crypto-bigint = { version = "0.4.8", features = ["serde"] }
+crypto-bigint = {version = "0.4.8", features = ["serde"]}
 
 # Logs
 log = "0.4.17"
@@ -65,20 +65,20 @@ slog-stdlog = "4.1.1"
 slog-term = "2.9.0"
 
 # Tracing
-tracing = { version = "0.1", features = [] }
-tracing-subscriber = "0.3"
+opentelemetry = {version = "0.18.0", features = ["rt-tokio"]}
+opentelemetry-datadog = {version = "0.6.0", features = ["reqwest-client"]}
+tracing = {version = "0.1", features = []}
+tracing-log = {version = "0.1", features = ["env_logger"]}
 tracing-opentelemetry = "0.18.0"
-#tracing-log = { version = "0.1", features = ["env_logger"] }
-opentelemetry = { version = "0.18.0", features = ["rt-tokio"] }
-opentelemetry-datadog = { version = "0.6.0", features = ["reqwest-client"] }
+tracing-subscriber = "0.3"
 
 # Errors
 anyhow = "1.0.57"
 thiserror = "1.0.31"
 
 # Serde
-serde = { version = "1.0.137", features = ["derive"] }
-serde_json = { version = "1.0.81" }
+serde = {version = "1.0.137", features = ["derive"]}
+serde_json = {version = "1.0.81"}
 
 # Datetime
 chrono = "0.4"
@@ -93,14 +93,14 @@ url = "2.2.2"
 mockall = "0.11.2"
 
 # Local dependecies
-marketplace-domain = { path = "../marketplace-domain" }
-marketplace-tests = { path = "../marketplace-tests" }
+marketplace-domain = {path = "../marketplace-domain"}
+marketplace-tests = {path = "../marketplace-tests"}
 
 [dev-dependencies]
 assert_matches = "1.5"
 claim = "0.5"
 envtestkit = "1.1.2"
 mockito = "0.31.0"
+rand = "0.8.5"
 rstest = "0.15.0"
 wiremock = "0.5"
-rand = "0.8.5"
