--- marketplace-infrastructure/src/logger/env.rs
@@ -1,6 +1,0 @@
-use slog::Drain;
-use slog_envlogger::EnvLogger;
-
-pub fn drain<D: Drain>(drain: D) -> EnvLogger<D> {
-	EnvLogger::new(drain)
-}
